Se crean los "exceptions" y se modifican url de postcontroller

+++ backend/posts_ms/src/main/java/com/Mintagram/posts_ms/controllers/PostController.java
@@ -1,5 +1,6 @@
 package com.Mintagram.posts_ms.controllers;
 
+import com.Mintagram.posts_ms.exceptions.PostNotFoundException;
 import com.Mintagram.posts_ms.models.Post;
 import com.Mintagram.posts_ms.repositories.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,40 +19,53 @@ public class PostController {
         this.postRepository = postRepository;
     }
 
+    /*@GetMapping("/")
+    String messageRoot(){
+        return "Conexion exitosa a PostMS";
+    }*/
+
+    @GetMapping("/")
+    String messagePostRoot(){
+        return "Por favor complete la url con el endpoint que corresponda";
+    }
+
+    @GetMapping("/all")
+    List<Post> getPosts(){
+        return postRepository.findAll();
+    }
+
+    @GetMapping("/byid/{id}")
+    Post getPostbyId(@PathVariable String id) {
+        Post post= postRepository.findById(id).orElse(null);
+        if (post==null)
+            throw new PostNotFoundException("No hay posts con id: " + id);
+        return post;
+    }
+
+    @GetMapping("/byuser/{username}")
+    List<Post> findPostsByUserName(@PathVariable String username){
+        return postRepository.findByUsername(username);
+    }
+
+    @GetMapping("/bycategoria/{categoria}")
+    List<Post> findPostsByCategoria(@PathVariable String categoria){
+        return postRepository.findByCategoria(categoria);
+    }
+
     @PostMapping("/create")
     Post createPost(@RequestBody Post post){
         post.setPostdate(new Date());
         return postRepository.save(post);
     }
 
-    @GetMapping("/all/")
-    List<Post> getPosts(){
-        return postRepository.findAll();
-    }
-
-    @GetMapping("/g/{id}")
-    Post getPosts(@PathVariable String id) {
-        return postRepository.findById(id).orElse(null);
-
-    }
-
-    @GetMapping("/{username}")
-    List<Post> findPostsByUserName(@PathVariable String username){
-
-        return postRepository.findByUsername (username);
-    }
-
-    @GetMapping("/get/{categoria}")
-    List<Post> findPostsByCategoria(@PathVariable String categoria){
-        return postRepository.findByCategoria(categoria);
-    }
-
     @PutMapping("/update")
     Post updatePostByComment(@RequestBody Post post){
-        Post post_update=postRepository.findById(post.getUsername()).orElse(null);
-                post_update.setDescription(post.getDescription());
-                post_update.setCategoria(post.getCategoria());
-                return postRepository.save(post_update);
+        Post post_update=postRepository.findById(post.getId()).orElse(null);
+        if (post_update==null)
+            throw new PostNotFoundException("No hay posts con id: " + post.getId());
+        post_update.setDescription(post.getDescription());
+        post_update.setCategoria(post.getCategoria());
+        return postRepository.save(post_update);
 
     }
 
@@ -59,8 +73,10 @@ public class PostController {
     @DeleteMapping("/remove/{id}")
     String deletePost(@PathVariable String id) {
         Post post = postRepository.findById(id).orElse(null);
+        if (post==null)
+            throw new PostNotFoundException("No hay posts con id: " + id);
         postRepository.deleteById(id);
-        return "borrado exitoso";
+        return "Borrado exitoso";
     }
 
 }
